Setup a proper neovim developer-environment
I've finally abandoned my organic nvim config that I've added to
whenever I needed a feature or stumbled upon a shortcoming. In favour of
a much more cohesive and well-thought environment like that of
ThePrimeagen.[^1]

I've wrapped my head around everything but a couple lines in `remap.lua`
and after that, all that's left to do is to familiarize myself with the
new keymaps and concepts.

As well as, finally learning how to use vim fluently.[^2] ;-;

[^1]: 0 to LSP : Neovim RC From Scratch - https://youtu.be/w7i4amO_zaE
[^2]: How to Vim in 2023: Tips and Tricks - https://youtu.be/FrMRyXtiJkc